requirements.txt
streamlit==1.27.2
Pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
numpy==1.26.2
requests==2.31.0
pandas==2.1.3